+++ src/test/resources/de/dfki/mary/LoadHelloIT.groovy
@@ -1,11 +1,11 @@
-package $project.marytts.component.packageName
+package my.world
 
 import marytts.server.MaryProperties
 import marytts.util.MaryRuntimeUtils
 
 import org.testng.annotations.*
 
-class Load${project.marytts.component.name}IT {
+class LoadHelloIT {
 
     @BeforeMethod
     void setup() {
@@ -15,13 +15,9 @@ class Load${project.marytts.component.name}IT {
     @DataProvider
     Object[][] properties() {
         [
-${project.marytts.component.config.collect { key, value ->
-    if (value instanceof List)
-        "            ['$key" + ".list', " + value.inspect() + "]"
-    else
-        "            ['$key', '$value']"
-    }.findAll {!it.contains('locale') }.join(',\n')
-}
+            ['foo.bar', 'foo baz'],
+            ['foo.qux.list', ['quux', 'quuux']],
+            ['foo.fnord', 'jar:/path/to/the/fnord']
         ]
     }
 
@@ -29,7 +25,7 @@ ${project.marytts.component.config.collect { key, value ->
     public void canGetProperty(name, expected) {
         def actual
         switch (name) {
-            case ~/.+\\.list\$/:
+            case ~/.+\.list$/:
                 actual = MaryProperties.getList(name)
                 assert actual.containsAll(expected)
                 break
@@ -38,7 +34,7 @@ ${project.marytts.component.config.collect { key, value ->
                 assert expected == actual
                 break
         }
-        if ("\$expected".startsWith('jar:')) {
+        if ("$expected".startsWith('jar:')) {
             assert MaryProperties.getStream(name)
         }
     }